Add more and beter tests for JDT signatures
 - Cover all 3 possibilities
 - Fix an issue with GenericUtilities
 - Take the chance to unify 2 different tests classes on the same parser
